refactor: move editor-version type and module
Move editor-version type into corresponding module and move that module into the types folder
