+++ src/types/editor-version.ts
@@ -1,4 +1,14 @@
-import { EditorVersion } from "../types/global";
+export type EditorVersion = {
+  major: number;
+  minor: number;
+  patch?: number;
+  flag?: "a" | "b" | "f" | "c";
+  flagValue?: 0 | 1 | 2;
+  build?: number;
+  loc?: string;
+  locValue?: number;
+  locBuild?: number;
+};
 
 /**
  * Compares two editor versions for ordering
